refactor(phase-3A): extract bfts_loop.py from cli/__init__.py [no behavior change]
Phase 3A continuation / cli §3.

Move the BFTS execution plumbing out of ``ari/cli/__init__.py``:

- ``ari/cli/bfts_loop.py`` (new, 769 lines) — owns
  ``_run_loop`` / ``_save_tree_incremental`` / ``_save_checkpoint``.
  Imports the lineage helpers from ``ari.cli.lineage`` and
  ``PathManager`` from ``ari.paths``.
- ``ari/cli/__init__.py`` shrinks to 1001 lines (was 1723) and now
  re-exports ``_save_tree_incremental`` / ``_run_loop`` /
  ``_save_checkpoint`` so external imports keep working.

Test fixture updates (cross-file source-text checks):
- ``test_idea_integration.py:test_run_loop_loads_idea_json_for_expand``
- ``test_server.py:test_cli_run_loop_max_workers_zero_safe``
Both now concatenate every ``ari/cli/*.py`` so the literal source
matches survive the split.

pytest ari-core/tests/ -q (excluding live integrations):
1971 passed, 15 skipped (unchanged).
